SAM3/common/requirements.txt
# Core
ultralytics==8.4.14

# PyTorch - install the variant matching your hardware:
#   CPU only : pip install torch torchvision
#   CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   CUDA 12.4: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch>=2.0.0
torchvision

# SAM3 text support (VERY IMPORTANT)
timm==1.0.24
git+https://github.com/ultralytics/CLIP.git

# Image & plotting
numpy==1.26.4
matplotlib
pillow
flask

# Video processing
opencv-python-headless>=4.10.0
